# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users":{"get":{"operationId":"UsersController_findAll","parameters":[{"name":"role","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Пользователи"],"security":[{"bearer":[]}]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Пользователи"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Пользователи"],"security":[{"bearer":[]}]}},"/api/auth/sign-up":{"post":{"operationId":"AuthController_signUp","summary":"Регистрация нового пользователя","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["Авторизация"]}},"/api/auth/sign-in":{"post":{"operationId":"AuthController_signIn","summary":"Авторизация пользователя","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["Авторизация"]}},"/api/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Авторизация"],"security":[{"bearer":[]}]}},"/api/auth/refresh":{"get":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":""}},"tags":["Авторизация"]}},"/api/account":{"get":{"operationId":"AccountController_profile","summary":"Получение профиля авторизованного пользователя","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDto"}}}}},"tags":["Аккаунт пользователя"],"security":[{"bearer":[]}]}},"/api/account/avatar":{"patch":{"operationId":"AccountController_updateAvatar","summary":"Обновление фотографии профиля","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":""}},"tags":["Аккаунт пользователя"],"security":[{"bearer":[]}]}},"/api/account/add-factory":{"post":{"operationId":"AccountController_addFactory","summary":"Добавление нового фабрики","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFactoryDto"}}}},"responses":{"200":{"description":""}},"tags":["Аккаунт пользователя"],"security":[{"bearer":[]}]}},"/api/account/factories":{"get":{"operationId":"AccountController_getFactories","summary":"Получение всех фабрик","parameters":[],"responses":{"200":{"description":""}},"tags":["Аккаунт пользователя"],"security":[{"bearer":[]}]}},"/api/cities":{"post":{"operationId":"CitiesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCityDto"}}}},"responses":{"201":{"description":""}},"tags":["Города"],"security":[{"bearer":[]}]},"get":{"operationId":"CitiesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Города"],"security":[{"bearer":[]}]}},"/api/cities/{id}":{"get":{"operationId":"CitiesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Города"],"security":[{"bearer":[]}]}},"/api/applications":{"post":{"operationId":"ApplicationsController_create","summary":"Создание новой заявки на регистрацию","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateApplicationDto"}}}},"responses":{"201":{"description":""}},"tags":["Заявки на регистрацию"]},"get":{"operationId":"ApplicationsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Заявки на регистрацию"],"security":[{"bearer":[]}]}},"/api/applications/{id}":{"get":{"operationId":"ApplicationsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Заявки на регистрацию"],"security":[{"bearer":[]}]}},"/api/applications/{id}/approve":{"patch":{"operationId":"ApplicationsController_approve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Заявки на регистрацию"],"security":[{"bearer":[]}]}},"/api/districts":{"post":{"operationId":"DistrictsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDistrictDto"}}}},"responses":{"201":{"description":""}},"tags":["Районы"],"security":[{"bearer":[]}]},"get":{"operationId":"DistrictsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Районы"],"security":[{"bearer":[]}]}},"/api/districts/{id}":{"get":{"operationId":"DistrictsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Районы"],"security":[{"bearer":[]}]},"patch":{"operationId":"DistrictsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDistrictDto"}}}},"responses":{"200":{"description":""}},"tags":["Районы"],"security":[{"bearer":[]}]},"delete":{"operationId":"DistrictsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Районы"],"security":[{"bearer":[]}]}},"/api/propertytypes":{"post":{"operationId":"PropertyTypesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePropertyTypeDto"}}}},"responses":{"201":{"description":""}},"tags":["Типы собственности"],"security":[{"bearer":[]}]},"get":{"operationId":"PropertyTypesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Типы собственности"],"security":[{"bearer":[]}]}},"/api/propertytypes/{id}":{"get":{"operationId":"PropertyTypesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Типы собственности"],"security":[{"bearer":[]}]},"patch":{"operationId":"PropertyTypesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePropertyTypeDto"}}}},"responses":{"200":{"description":""}},"tags":["Типы собственности"],"security":[{"bearer":[]}]},"delete":{"operationId":"PropertyTypesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Типы собственности"],"security":[{"bearer":[]}]}},"/api/units":{"post":{"operationId":"UnitsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUnitDto"}}}},"responses":{"201":{"description":""}},"tags":["Ед. измерения"],"security":[{"bearer":[]}]},"get":{"operationId":"UnitsController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Ед. измерения"],"security":[{"bearer":[]}]}},"/api/units/{id}":{"get":{"operationId":"UnitsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Ед. измерения"],"security":[{"bearer":[]}]},"patch":{"operationId":"UnitsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUnitDto"}}}},"responses":{"200":{"description":""}},"tags":["Ед. измерения"],"security":[{"bearer":[]}]},"delete":{"operationId":"UnitsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Ед. измерения"],"security":[{"bearer":[]}]}},"/api/packages":{"post":{"operationId":"PackagesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePackageDto"}}}},"responses":{"201":{"description":""}},"tags":["Тип фасовки"],"security":[{"bearer":[]}]},"get":{"operationId":"PackagesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Тип фасовки"],"security":[{"bearer":[]}]}},"/api/packages/{id}":{"get":{"operationId":"PackagesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Тип фасовки"],"security":[{"bearer":[]}]},"patch":{"operationId":"PackagesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePackageDto"}}}},"responses":{"200":{"description":""}},"tags":["Тип фасовки"],"security":[{"bearer":[]}]},"delete":{"operationId":"PackagesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Тип фасовки"],"security":[{"bearer":[]}]}},"/api/confectionarytypes":{"post":{"operationId":"ConfectionarytypesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateConfectionaryTypeDto"}}}},"responses":{"201":{"description":""}},"tags":["Тип сладости"],"security":[{"bearer":[]}]},"get":{"operationId":"ConfectionarytypesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Тип сладости"],"security":[{"bearer":[]}]}},"/api/confectionarytypes/{id}":{"get":{"operationId":"ConfectionarytypesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Тип сладости"],"security":[{"bearer":[]}]},"patch":{"operationId":"ConfectionarytypesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConfectionarytypeDto"}}}},"responses":{"200":{"description":""}},"tags":["Тип сладости"],"security":[{"bearer":[]}]},"delete":{"operationId":"ConfectionarytypesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Тип сладости"],"security":[{"bearer":[]}]}}},"info":{"title":"SLADKOEZHKOVO API","description":"API к курсовому проекту по ПСБД <br/> &copy; Evteev Dmitry 2024","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"Bearer","bearerFormat":"JWT","type":"http","in":"Header"}},"schemas":{"UpdateUserDto":{"type":"object","properties":{"handle":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"lastName":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"role":{"type":"string"}}},"SignUpDto":{"type":"object","properties":{"handle":{"type":"string"},"email":{"type":"string"},"lastName":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}},"required":["handle","email","lastName","firstName","middleName","password","role"]},"SignInDto":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}},"required":["login","password"]},"ErrorDto":{"type":"object","properties":{"message":{"type":"string"},"error":{"type":"string"},"statusCode":{"type":"number"}},"required":["message","error","statusCode"]},"ProfileDto":{"type":"object","properties":{"id":{"type":"string"},"handle":{"type":"string"},"email":{"type":"string"},"lastName":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"imageLink":{"type":"string"},"role":{"type":"string"}},"required":["id","handle","email","lastName","firstName","middleName","imageLink","role"]},"CreateFactoryDto":{"type":"object","properties":{"name":{"type":"string"},"handle":{"type":"string"},"cityId":{"type":"string"},"phoneNumber":{"type":"string"},"propertyTypeId":{"type":"string"},"year":{"type":"number"}},"required":["name","handle","cityId","phoneNumber","propertyTypeId","year"]},"CreateCityDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"CreateApplicationDto":{"type":"object","properties":{"role":{"type":"string"},"lastName":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"email":{"type":"string"},"handle":{"type":"string"}},"required":["role","lastName","firstName","middleName","email","handle"]},"CreateDistrictDto":{"type":"object","properties":{"name":{"type":"string"},"cityId":{"type":"string"}},"required":["name","cityId"]},"UpdateDistrictDto":{"type":"object","properties":{"name":{"type":"string"},"cityId":{"type":"string"}}},"CreatePropertyTypeDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdatePropertyTypeDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreateUnitDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdateUnitDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreatePackageDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdatePackageDto":{"type":"object","properties":{"name":{"type":"string"}}},"CreateConfectionaryTypeDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"UpdateConfectionarytypeDto":{"type":"object","properties":{"name":{"type":"string"}}}}}}